+++ apps/next/pages/dashboard/search/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { SearchScreen } from "app/features/dashboard/search";
-import { AuthGuard } from "app/utils/auth-guard";
+import { AuthGuard } from "app/utils/authGuard";
 
 export default function SearchPage() {
   return (
