+++ packages/app/design-system/index.tsx
@@ -2,6 +2,6 @@ export { View } from "app/design-system/view";
 export { Text } from "app/design-system/text";
 export { TextLink } from "app/design-system/text-link";
 export { Pressable } from "app/design-system/pressable";
-export { TextInput } from "app/design-system/text-input";
-export { ActivityIndicator } from "app/design-system/activity-indicator";
+export { TextInput } from "app/design-system/textInput";
+export { ActivityIndicator } from "app/design-system/activityIndicator";
 export { Button } from "app/design-system/button";
